healthcare-api/app/schemas/user_schema.py
from pydantic import BaseModel, EmailStr, Field


class UserCreate(BaseModel):
    name: str = Field(min_length=2, max_length=100)
    email: EmailStr
    password: str = Field(min_length=6)
    role: str = "patient"


class UserOut(BaseModel):
    id: str
    name: str
    email: EmailStr
    role: str
    is_approved: bool
