healthcare-api/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
email-validator
python-dotenv
